aotriton_0_11: init at 0.11b
This change adds aotriton_0_11 and uses it for Torch 2.9 builds.

This PR also changes the argument handling for ROCm clang/lld. aotriton
uses a unicode character (＊) in its generated source files. The Nix
compiler wrapper does not correctly handle this when using a compiler
response file. For this reason, we disabled the use of response files
in #71. However, this bites us now with aotriton 0.11b -- the
compiler/linker is passed so many arguments (more kernels + an extra
arch) that we hit the argument length limit.

To be able to build aotriton 0.11b, we have to re-enable use of
compiler/linker response files, so that we can pass the arguments in a
file to circumvent the argument limit. So we vendor `cc-wrapper.sh` from
nixpkgs and fix the root cause, which was the use of shell quoting
(`printf "%q\n"`) when writing the response file. In the C locale, this
leads to re-encoding non-ASCII characters as octal numbers. Instead, we
just echo the arguments to the file.