Update nixpkgs (#130)
* Update nixpkgs

* Switch to torch-bin as the default Torch

* Disable einops override?

* ucx: fix build

* aotriton: remove msgpack dependency

* The triton-rocm package is not needed anymore

All the changes (like patching up the linker path) are in the main
triton derivation.

* cudaPackages: remove

Let's try to hop on the upstream CUDA packages again. We started
diverging a lot (e.g. upstream added "include" outputs to the CUDA
packages) and we do not need CUDA 11.8 anymore since dropping
Torch 2.7.

* python3Packages.triton-xpu: relax cmake bound

* gcc: fix a warning

* Remove old gcc versions

* nix fmt

* rocmPackages.aotriton_0_{10,11}: sync and speed up build

Sync the derivation (mostly) with nixpkgs. Use pre-compiled kernel
images to speed up the build.

* python3Packages.torch-bin: correctly remove aotriton

* python3Packages.transformers: 4.55.0 -> 4.57.1

Needed for compatibility with tokenizers.